+++ fork/AlchemyStars/src/Alchemist.UI/AboutWindow.xaml.cs
@@ -7,7 +7,7 @@ namespace Alchemist.UI;
 
 public partial class AboutWindow : Window
 {
-    public string AppVersion { get; } = typeof(AboutWindow).Assembly.GetName().Version?.ToString(3) ?? "1.1.6";
+    public string AppVersion { get; } = typeof(AboutWindow).Assembly.GetName().Version?.ToString(3) ?? "1.1.7";
     public string RuntimeVersion { get; } = RuntimeInformation.FrameworkDescription;
     public string OperatingSystem { get; } = RuntimeInformation.OSDescription;